refactor: extract Banner component to new file
