+++ src/views/components/GamePanel/GamePanel.tsx
@@ -1,7 +1,7 @@
-import { Game, ImageCommons } from '@data/types'
-import { ScreenshotsList } from '@views/components'
+import { Game } from '@data/types'
+import { Banner } from '@views/components'
 import * as React from 'react'
-import { ComponentProps, PropsWithChildren, useState } from 'react'
+import { PropsWithChildren } from 'react'
 import style from './GamePanel.module.scss'
 import { IGamePanel } from './IGamePanel'
 import { StylingUtils } from '@src/utils'
@@ -51,42 +51,6 @@ function Header({ game }: IHeader) {
 
 			<ul className={style.Tags}>{gameGenres}</ul>
 		</header>
-	)
-}
-
-interface IBanner extends ComponentProps<'article'> {
-	name: string
-	images: ImageCommons[]
-}
-
-function Banner(props: IBanner) {
-	const { name, images } = props
-
-	const [currentIndex, setCurrentIndex] = useState<number>(0)
-
-	const selectImage = (e: React.MouseEvent<HTMLElement>) => {
-		const target   = e.target as HTMLElement
-		const listItem = target.closest('[data-index]') as HTMLLIElement
-
-		if (! listItem) return
-
-		const parsedIndex = listItem.dataset['index'] as unknown
-		setCurrentIndex(parsedIndex as number)
-	}
-
-	return (
-		<article className={style.Banner}>
-			<figure className={style.Main}>
-				<img
-					src={images[currentIndex]?.image ?? ''}
-					alt={`main banner for the "${name}" game`}
-				/>
-			</figure>
-
-			<menu className={style.Screenshots} onClick={selectImage}>
-				<ScreenshotsList {...props} />
-			</menu>
-		</article>
 	)
 }
